requirements.txt
Flask>=2.0.0
opencv-python>=4.8.0
pyzbar>=0.1.9
numpy>=1.24.0
Pillow>=10.0.0
